background task only runs once and loads sites, if they are available #21
